Bind-mount /dev/pts, /dev/shm, and /sys/fs/fuse/connections.
This should fix crosh not being able to launch, as well as be the proper fix for
Chromium/Chrome not launching in the chroot. This may preemptively fix issues
with mounting removable media from inside the chroot, but this has not been
tested, so be wary of data corruption on the removale media.