Add powerd-poking daemon for massive power savings.
The powerd daemon auto-launches with X11 and pokes the system powerd while the
X11 session is not idle. It also monitors the screensaver setting and
continuously pokes powerd if the screensaver is inhibited (e.g. you are watching
a video).

This means that we can leave powerd running, which has huge battery benefits.