Override powerd_suspend persist if running a chroot from removable media.
powerd_suspend usually disables USB device persist on suspend, since Chromium OS
is designed to handle it and it saves a half a second during suspend and resume.
If you're using a chroot on an SD card or USB stick, however this will cause the
chroot to break when you suspend, which is generally considered a bad thing.

This modifies powerd_suspend on the fly to not disable persist on USB devices.
It establishes an overlay on /usr/bin, since it is normally read-only. The
overlay is removed in a best-effort manner when there are no more external-media
chroots mounted.