Restore tap-to-click when switching to the chroot with Xephyr.
When typing, Chromium OS disables tap-to-click. This normally gets re-enabled
once the mouse moves, but when launching or switching to a chroot, Aura does not
get any more mouse movement events and so cannot re-enable tap-to-click. We now
do so manually upon window switch, for an arbitrary number of trackpads.

Fixes #38.