Change handling of USB persistence (fixes #288)
In the old way of doing things ChromeOS would turn off device
persistence to save a few milliseconds of suspend time. Crouton had to
jump though hoops to stop it doing this if the chroot was on a USB
device.

Since upstream merged commit: 0007a546853a529064772b1b96bd9164dece8c46
into power_manager.git and defaulted the kernel to have USB persistence
off all we now need to do is turn it on when we detect a chroot in
/media (where USB devices are mounted).

Arguably you could be smarter about which USB devices to enable but at
least by doing them all you ensure nothing breaks and it works.