mount chrootfs read-only at ~/Downloads/crouton-*
This patch bind-mounts $CHROOT read-only into ~/Downloads/crouton-$NAME
if we're logged in as a user so that we can access files in the chroot
directly from Chromium OS' "Files" application (e.g. to upload files).

In addition, mounting read-only protects us from the danger of Chromium
OS deleting files randomly from ~/Downloads whenever disk space is low:

  https://support.google.com/chromebook/answer/1061547

In fact, even the root user cannot delete files from or modify files
within a read-only mount without first remounting it as read-write!

CAVEAT: I'm assuming that Chromium OS is not so aggressive in freeing up
disk space that it would find and remount this bind-mount as read-write.