enter-chroot: Simply bind mount /var/run/udev
Also invert the test (i.e. check for existence of /var/run/udev),
to protect against the very unfortunate case where /var/run/udev
is a symlink to /dev/.udev (update while chroot is running, started
with previous version of enter-chroot). In that case, bindmount
creates /dev/.udev, which is not a problem in itself, but would
break further chroot starts.