croutonclip: Refactor triggering code
For some reason bash 4.3.11 (trusty) segfaults often with the
previous version of the code, presumably because traps/sigaction
are not reset properly in children.

We refactor the code to receive signals in a subshell, and print
a line everytime a signal is received. We then read these lines
from the other end of a pipe, and transfer the clipboard.

Tested:
 - arch bash 4.3.24
 - trusty bash 4.3.11 / dash
 - precise bash 4.2.25 / dash (Xephyr)