Clarify creating a chroot with a bootstrap tarball
Clarify that you also need to specify the `-t` parameter after downloading the bootstrap files.

Given that you don't get an error about missing a target when running `-f` without `-t`, it took me quite a while and some googling to figure out why the command in step 4 didn't work; given that it was in the readme and every other command before this didn't need additional parameters, it was surprising that I needed to also add the `-t` parameter. I kind of get the distinction between a bootstrap and a target now that I've done it but I think mentioning it will save someone else from having to figure out what they are doing wrong after literally following the steps in the readme.