websocket.h: Refactor popen2, do not leak fds.
Close the unneeded ends of the pipes in the parent to prevent fd
leakage. Fixes #1286.

This has the advantage of throwing a POLLHUP event when the child
closes its stdout. Refactor popen2 accordingly: Do all
input/output first, then check exit status. This allows
success/error paths to share more code.

Also, return an error if not all data can be written; and
waitpid(pid, NULL, 10) never worked (the 3rd parameter is not a
timeout).

Test code:
```
// gcc -Wall test.c && (ulimit -n 16; ./a.out)

#include "websocket.h"

int main() {
    char buffer[BUFFERSIZE];
    int i, status;
    verbose = 4;

    for (i = 0; i < 16; i++) {
        memset(buffer, 0, BUFFERSIZE);
        status = popen2("sha1sum", NULL, "1234", 4, buffer, BUFFERSIZE);
        trueorabort(status >= 40, "Answer too short");
        trueorabort(!strncmp(buffer,
                             "7110eda4d09e062aa5e4a390b0a572ac0d2c0220", 40),
                    "Invalid sha1sum");
        status = popen2("dummy1sum", NULL, "1234", 4, buffer, BUFFERSIZE);
        trueorabort(status < 0, "popen2 did not fail");
    }

    return 0;
}
```