fbserver-proto: Pass X11 keycode instead of X11 keysym
Instead of transmitting a X11 key symbol (e.g. letter 'A'), which
depends on the keyboard layout used in Chromium OS, pass a key
code (i.e. position on the keyboard), and let X11 translate it
to the appropriate symbol, depending on the keyboard layout in
the chroot.

We keep old structure in fbserver-proto.h, for backward
compatibility.