Nudge display brightness on return to cros
For some targets, when croutoncycling back to cros, the display brightness is not restored correctly - the host cros dbus server knows the correct value but the hardware does not (observed on Toshiba B3340). This admittedly hacky change nudges the the display brightness so that the old correct value is restored.