x11-common: divert /etc/upstart-xsessions
On xenial we will only ever pull in upstart as a dependency
of unity, but this means a simple echo > /etc/upstart-xsessions
will leave an empty file that will conflict when upstart is
installed with the unity target. Using dpkg-divert will also
behave properly when upstart is upgraded or reinstalled after
purging.