croutonxinitrc-wrapper: xorg target: kick the display on startup
Since Chromium 56.0.2923.0, Chromium tries to switch off the display when
switching VT (crbug.com/655770). For some unclear reason, running xrandr
forces the display to be back on, and this is not needed ever again
when switching VTs.
I'm not clear if the sleep is required, but since this is a race, this
should be a little bit safer.