Fix locale for Ubuntu Xenial
Previously, core checked for the absence of /etc/default/locale to
determine if a locale needed to be generated an set as the default.
Unfortunately, in Xenial /etc/default/locale exists by default in
a clean installation with a comment that reads:

    # File generated by update-locale

This commit enhances the test for /etc/default/locale to check that
the file contains `LANG=`, ensuring that a default locale is set.

Fixes: #2921 and #3283