Always show the logout menu in gnome
Gnome by default does not display the user section which contains the
logout option in the system menu when there is only one user. This is
a problem because there is normally only one user in a crouton chroot,
and there is no other way for the user to exit the chroot from the UI
(though it can be done using the CLI.) In this commit I install
dconf-tools and specifies a new system wide default to always show the
logout option.

I also incidentally separated the gnome packages in new lines and
alphabetized them for readability.

Fixes: #3243