Poke the new usage stat file
This helps give the ChromeOS team a vague view of the approximate number of active crouton machines (who have opted into ChromeOS stats reporting).  This single-bit stat is not associated with users, machines, or the actual usage of crouton.  If this terrifies you, you can block the stat by setting CROUTON_DISABLE_STATS in your environment (the variable is plural but there are no other stats).

https://crbug.com/989219 for reference