Wrap uname to enable glibc installs for kernel revisions >= 255
Fixes #4429, although we're not yet patching uname syscalls.
The key issue is that glibc's version arithmetic can result it in
interpreting revision>=255 as minor+1, which could in turn result in
glibc trying to use kernel features that don't actually exist.
If we find that applications are crashing with SIGSYS then we'll also
need an LD_PRELOAD hack like the one here:
https://gist.github.com/dnschneid/5c88a1802868c1be74d4debb5627c8c7